Finalisations routes products

+++ products/src/main/kotlin/iut/nantes/project/products/services/FamilleService.kt
@@ -1,17 +1,22 @@
 package iut.nantes.project.products.services
 
 import iut.nantes.project.products.controllers.dto.FamilleDto
+import iut.nantes.project.products.exceptions.FamilleHasLinkedProductsException
 import iut.nantes.project.products.exceptions.FamilleNameConflictException
 import iut.nantes.project.products.exceptions.FamilleNotFoundException
 import iut.nantes.project.products.exceptions.InvalidIdFormatException
 import iut.nantes.project.products.repositories.FamilleRepository
+import iut.nantes.project.products.repositories.ProductRepository
 import org.springframework.core.env.Environment
 import org.springframework.dao.DataIntegrityViolationException
 import java.lang.IllegalArgumentException
 import java.util.UUID
 
 class FamilleService(
-    private val familleRepository: FamilleRepository, private val environment: Environment
+    private val familleRepository: FamilleRepository,
+    private val productRepository: ProductRepository,
+    private val environment: Environment,
+
 ) {
     fun createFamille(famille: FamilleDto): FamilleDto {
         if (!environment.activeProfiles.contains("test")) famille.id = UUID.randomUUID().toString()
@@ -56,15 +61,16 @@ class FamilleService(
     }
 
     fun deleteFamille(id: String) {
+
         val family = familleRepository.findById(id)
             .orElseThrow { FamilleNotFoundException("Famille with ID $id not found.") }
 
         // Vérification s'il y a des produits liés à cette famille
-        //if (family.products.isNotEmpty()) {
-        //    throw FamilyHasLinkedProductsException("Cannot delete a family with linked products.")
-        //}
+        if (productRepository.findAll().any{ it.family.id == id}) {
+            throw FamilleHasLinkedProductsException("Cannot delete a family with linked products.")
+        }else {
+            familleRepository.deleteById(id)
+        }
 
-
-        familleRepository.deleteById(id)
     }
 }
